Rename STAT -> StatBuilder, add build() fn
This is a refactor/cleanup, no functional change.
