+++ fea-rs/src/compile/output.rs
@@ -1,17 +1,13 @@
 //! The result of a compilation
 
-use std::collections::{BTreeMap, HashMap};
+use std::collections::BTreeMap;
 
 use write_fonts::{
     dump_table,
     read::{FontRef, TableProvider},
-    tables::{
-        self,
-        layout::{FeatureParams, SizeParams},
-        stat::AxisValueRecord,
-    },
+    tables::layout::{FeatureParams, SizeParams},
     types::Tag,
-    FontBuilder, OffsetMarker,
+    FontBuilder,
 };
 
 use super::{
@@ -19,10 +15,7 @@ use super::{
     lookups::{AllLookups, FeatureKey, LookupId},
     tables::{NameSpec, Tables},
 };
-use crate::{
-    compile::tables::{AxisLocation, StatFallbackName},
-    Diagnostic,
-};
+use crate::Diagnostic;
 
 /// The output of a compilation operation.
 ///
@@ -87,79 +80,9 @@ impl Compilation {
         //});
 
         let mut name_builder = self.tables.name.clone();
-        if let Some(stat_raw) = self.tables.STAT.as_ref() {
-            let elided_fallback_name_id = match &stat_raw.name {
-                StatFallbackName::Id(id) if name_builder.contains_id(*id) => *id,
-                StatFallbackName::Id(id) => {
-                    panic!("ElidedFallbackNameID '{}' does not exist in font", id)
-                }
-                StatFallbackName::Record(names) => name_builder.add_anon_group(&names),
-            };
-
-            let mut design_axes = Vec::new();
-            for record in &stat_raw.records {
-                let name_id = name_builder.add_anon_group(&record.name);
-                design_axes.push(tables::stat::AxisRecord {
-                    axis_tag: record.tag,
-                    axis_name_id: name_id,
-                    axis_ordering: record.ordering,
-                });
-            }
-            design_axes.sort_unstable_by_key(|x| x.axis_tag);
-            let axis_indices = design_axes
-                .iter()
-                .enumerate()
-                .map(|(i, val)| (val.axis_tag, i as u16))
-                .collect::<HashMap<_, _>>();
-
-            let axis_values: Vec<OffsetMarker<tables::stat::AxisValue>> = stat_raw
-                .values
-                .iter()
-                .map(|axis_value| {
-                    let flags =
-                        tables::stat::AxisValueTableFlags::from_bits(axis_value.flags).unwrap();
-                    let name_id = name_builder.add_anon_group(&axis_value.name);
-                    let to_add = match &axis_value.location {
-                        AxisLocation::One { tag, value } => tables::stat::AxisValue::format_1(
-                            //TODO: validate that all referenced tags refer to existing axes
-                            *axis_indices.get(tag).unwrap(),
-                            flags,
-                            name_id,
-                            *value,
-                        ),
-                        AxisLocation::Two {
-                            tag,
-                            nominal,
-                            min,
-                            max,
-                        } => {
-                            let axis_index = *axis_indices.get(tag).unwrap();
-                            tables::stat::AxisValue::format_2(
-                                axis_index, flags, name_id, *nominal, *min, *max,
-                            )
-                        }
-                        AxisLocation::Three { tag, value, linked } => {
-                            let axis_index = *axis_indices.get(tag).unwrap();
-                            tables::stat::AxisValue::format_3(
-                                axis_index, flags, name_id, *value, *linked,
-                            )
-                        }
-                        AxisLocation::Four(values) => {
-                            let mapping = values
-                                .iter()
-                                .map(|(tag, value)| {
-                                    AxisValueRecord::new(*axis_indices.get(tag).unwrap(), *value)
-                                })
-                                .collect();
-                            tables::stat::AxisValue::format_4(flags, name_id, mapping)
-                        }
-                    };
-                    OffsetMarker::new(to_add)
-                })
-                .collect::<Vec<_>>();
-            let mut table = tables::stat::Stat::new(design_axes, axis_values.into());
-            table.elided_fallback_name_id = Some(elided_fallback_name_id);
-            builder.add_table(Tag::new(b"STAT"), dump_table(&table).unwrap());
+        if let Some(stat_raw) = self.tables.stat.as_ref() {
+            let stat = stat_raw.build(&mut name_builder);
+            builder.add_table(Tag::new(b"STAT"), dump_table(&stat).unwrap());
         }
 
         let (gsub, mut gpos) = self.lookups.build(&self.features);
